+++ examples/ai-core/src/stream-text/openai-reasoning-model.ts
@@ -17,8 +17,7 @@ async function main() {
   console.log();
   console.log('Usage:', {
     ...(await result.usage),
-    reasoningTokens: (await result.experimental_providerMetadata)?.openai
-      ?.reasoningTokens,
+    reasoningTokens: (await result.providerMetadata)?.openai?.reasoningTokens,
   });
   console.log('Warnings:', await result.warnings);
 }