+++ examples/ai-core/src/stream-text/perplexity.ts
@@ -19,7 +19,7 @@ async function main() {
   console.log('Finish reason:', await result.finishReason);
   console.log(
     'Metadata:',
-    JSON.stringify(await result.experimental_providerMetadata, null, 2),
+    JSON.stringify(await result.providerMetadata, null, 2),
   );
 }