+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- Updated dependencies [dc49119]
+  - ai@4.1.34
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [74f0f0e]
+  - ai@4.1.33
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [c128ca5]
   - ai@4.1.32